Change Folder Name
Logisti_Regression to Logistic_Regression (add c in logistic)
