+++ 01_Basic/Logistic_Regression/ver_jax.py
